microservices/notification_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
alembic
python-dotenv
pytest
httpx
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]
pydantic[email]
redis